SONARJAVA-5138 Override ProblemHandler, ProblemReporter and ASTParser
Added with ECJ 3.41: Prevent ASTParser to throw "Missing system library" when Java >=9 and "java.base" module is not in the classpath. (e.g. android context)

Added with ECJ 3.41: Fix sanity tests in sonar-java failing due to parser errors on non-compiling sources with the message "(...) is wrongly tagged as containing missing types".